requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
deep-translator